requirements.txt
python-telegram-bot>=20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
